Squashed commit:
    Add basic functionality of height maps.
    Break MeshRenderer into a SimpleMeshRenderer and a ComplexMeshRenderer.
    Triangulate and render height maps.
    Integrate edit height maps menu into spawn pt menu.
    Fix hard-coded developer paths in several writers.
    Fix problem that height map failed to load when created.
    Clean up usages of Error and Exception.
    Implement switching between bricks edit mode and height map edit mode.
    Break EditorStatusBar into EditorHUD and EditorState.
    Implement a variant of the cursor to be shown when editing a mesh.
    Rename Selection as BrickSelection. Move files among packages.
    Implement HeightMapSelection. Move input handling stuff into KeyboardHandler.
    Replace diamond cursor with a small variant of the bricks cursor.
    Implement keyboard actions to change height of selected heightmap samples.
    Implement saving heightmap along with scene.
    Implement resizing and resampling of a heightmap.
    Implement smooth shading of height maps.
    Remove MyDataOutputStream, use KDataOutputStream from kutils instead.
    Remove MyDataInputStream, use KDataInputStream from kutils instead.
    Migrate Groovy build scripts to Kotlin.
    Move versions to gradle/libs.versions.toml.
    Use layout.projectDirectory instead of deprecated buildDir.
    Rewrite generateSources task to get it to work with Gradle's config cache.
    Make BuildConfig a Kotlin object rather than a Java class.
    Update to Kotlin 2.3.0, and fix new warnings.

+++ import_asset/src/java/ch/digorydoo/titanium/import_asset/brick_textures/BrickTextureImporter.kt
@@ -147,7 +147,7 @@ class BrickTextureImporter(private val options: Options) {
 
         parts.forEachIndexed { i, part ->
             val x = padding + (i % arrangeAcross) * (partWidth + 2 * padding)
-            val y = padding + (i / arrangeAcross).toInt() * (partHeight + 2 * padding)
+            val y = padding + (i / arrangeAcross) * (partHeight + 2 * padding)
             dst.drawImage(part, x, y)
 
             if (padding > 0) {
